Add files via upload
Changed hashCode( ) calculation (added a part of the string array, but excluded *31 part), separated to two different classes.
